+++ client/src/App.tsx
@@ -5,7 +5,6 @@ import { pdfjs } from "react-pdf";
 import { largeSmall, useAppStateValue, useAsyncStateMachine } from "./State";
 import { QuestionPanel } from "./QuestionPanel";
 import { AnswerPanel } from "./AnswerPanel";
-import { ReviewPanel } from "./ReviewPanel";
 
 pdfjs.GlobalWorkerOptions.workerSrc = new URL(
   "pdfjs-dist/build/pdf.worker.min.js",
@@ -19,7 +18,6 @@ function App() {
   return (
     <div id="app" className={`question-${largeSmall(largeQuestionPanel)} review-${largeSmall(largeReviewPanel)} answer-${largeSmall(largeAnswerPanel)}`}>
       <QuestionPanel />
-      <ReviewPanel />
       <AnswerPanel />
     </div>
   )
